% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kkr_fit.R
\name{krr_fit}
\alias{krr_fit}
\title{KRR 모델 적합 함수}
\usage{
krr_fit(X, y, rho = 1, lambda = 1e-04)
}
\arguments{
\item{X}{입력 데이터 (matrix)}

\item{y}{반응값 (vector)}

\item{rho}{커널 파라미터}

\item{lambda}{페널티 파라미터}
}
\value{
krr 객체 (coefficients, fitted_values 등을 포함)
}
\description{
Gaussian kernel을 사용해 Kernel Ridge Regression(KRR)을 적합합니다.
}
