% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kkr_predict.R
\name{krr_predict}
\alias{krr_predict}
\title{KRR 예측 함수}
\usage{
krr_predict(object, X_new, ...)
}
\arguments{
\item{object}{krr 객체}

\item{X_new}{새로운 입력값 (matrix)}
}
\value{
예측값 벡터
}
\description{
krr_fit()으로 적합된 KRR 모델을 이용해 새로운 데이터의 예측값을 계산합니다.
}
